requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
numpy
pandas
scikit-learn==0.22
joblib
memoized-property
mlflow
gcsfs
google-cloud-storage
fastapi
uvicorn
